for loops

+++ packages/contracts/contracts/protocol/libraries/types/DistributionTypes.sol
@@ -29,7 +29,7 @@ library DistributionTypes {
    **/
   struct IncentivizedAsset {
     mapping(address => Reward) rewardData;
-    mapping(uint128 => address) rewardList;
+    mapping(uint256 => address) rewardList;
     uint128 numRewards;
     uint8 decimals;
   }
